Fix maximum argument buffer sizes
Summary:
For whatever reason, these may have changed on a recent version of macOS.

In order to fix this, I'm going to get the actual values here. The max number of pids doesn't appear to be available through a syscall, but I got a more reasonable value by fetching this from the current version of macOS and bumping it 2X.

We initialize the ARGMAX that we can at class load time and set this value appropriately. This should always succeed

Differential Revision: D23134077

fbshipit-source-id: 1d8d385cee6f9a2aaf96800d622a110e4f0fde9a