Use CoreSimulator API for settings approval
Summary: Where relevant we can use the new CoreSimulator APIs for dealing with approvals. This means that we can avoid some of the rather janky aspects of updating the TCC database and just call out to the relevant API. I expect that CoreSimulator is doing something similar under the hood, but without the need to restart springboard or any of that.

Reviewed By: c-ryan747

Differential Revision: D23160622

fbshipit-source-id: 37e1a7e5ede8323a2960a783ee41405cfb8a9f5c