add FBXCTestResultBundleParser
Summary: Previously, all the codes related to result bundle parsing lived in `FBDeviceXCTestCommands` because we could not get result bundle for simulator tests. I'm working to use `xcodebuild` instead of `XCTestBootstrap` to run the `idb xctest` command when users want to get the result bundle. The result bundle parsing and reporting codes could be shared for device and simulator xctests. It makes sense to extract and reuse the codes and put them under `XCTestBootstrap` where `FBXcodeBuildOperation` and `FBXCTestResultToolOperation` live.

Reviewed By: lawrencelomax

Differential Revision: D23184363

fbshipit-source-id: 975fa514728ac9fa672bf4c0f964d3e408ca5258