move a few methods from FBDeviceXCTestCommands to FBXcodeBuildOperation
Summary:
One is to get the xcodebuild path. The other is to create a xctestrun file from a test launch configuration. Both of them are related to xcodebuild so I think it makes sense to move them under `FBXcodeBuildOperation`. This way, they can be shared between device and simulator xctest.

Also changed `FBDeviceXCTestCommands` to use those new methods.

Reviewed By: lawrencelomax

Differential Revision: D23207037

fbshipit-source-id: 37d066814223ecd992d1b5fae9fea373f75331b5