use xcodebuild for simulator tests to get result bundle
Summary:
Added a `shouldUseXcodebuild` property in `FBTestLaunchConfiguration` to indicate whether we should use `xcodebuild` to run the xctest. We will set `shouldUseXcodebuild` if we install the test bundle with a .xctestrun file to follow the existing logic.

The reason is that if we install the test bundle with a .xctest file, we will get a `FBXCTestBootstrapDescriptor` descriptor. If we install the test bundle with a .xctestrun file, we will get a `FBXCodebuildTestRunDescriptor` descriptor: check the `FBIDBStorageManager.m` file for more details.

And then, we will get a `FBTestLaunchConfiguration` instance from a test descriptor (`FBXCTestBootstrapDescriptor` or `FBXCodebuildTestRunDescriptor`) with the `- (FBTestLaunchConfiguration *)testConfigWithRunRequest:(FBXCTestRunRequest *)request testApps:(FBTestApplicationsPair *)testApps logger:(id<FBControlCoreLogger>)logger;` instance method. So it makes sense to set `shouldUseXcodebuild` in `FBTestLaunchConfiguration` for a `FBXCodebuildTestRunDescriptor` descriptor that was resulted from the fact that the result bundle was installed with a .xctestrun file.

I think the existing logic that we will get different types of FBXCTestDescriptor based on how we install the test bundle is not very flexible. We might need to change that in the future if we want the ability to specify to use XCBootstrap or xcodebuild to run the tests no matter how we install the test bundle.

Reviewed By: lawrencelomax

Differential Revision: D23184417

fbshipit-source-id: e2f0d216fd3bc94c800b7514a4564dc8adccaad9