Add Provisioning Profile API
Summary: Adds a means of fetching provisioning profiles for devices. This is done by making a file-like overlay on top of Provisioning Profiles. This means that we don't need new APIs for the various related commands. Instead we can have sugar over this on the client, or just use it directly.

Reviewed By: fejese

Differential Revision: D22117888

fbshipit-source-id: 412c9d386e165e6d69d2b1df4d6663c87d8c0a7b