Add flags for file container types in CLI
Summary:
Adds container types so we can now support the media/root/crashes directory.

We have to send the old and the new sets of data, which the backend can handle both the old and the new parts of the protocol. After some time we can remove the deprecated parts on client and server as the `FileContainer` will take over here.

Reviewed By: c-ryan747

Differential Revision: D23103129

fbshipit-source-id: f4b2439b396422cd7e1bf2e1241774743b5aee45