Add FBDeviceActivationCommands
Summary:
Adds activation implementations. This is based upon the activation service that is in recent iOS's (I've tested 12 and above).

This will allow us to erase unactivated devices. To do this there's some token exchange back and forth between the device and two of Apple's endpoints. This is done in the companion and apes what cfgutil does.

Reviewed By: fejese

Differential Revision: D23595523

fbshipit-source-id: 884bc739a75c3b5acbfa8b8dd73f67b0a11d6933