Tighten Metadata and Failures in IdbClient.build
Summary:
This is extremely useful, since we'll automatically populate this information. It will both pass and pull information with the given arguments provided.

Additionally, we improve the exception messages to output the appropriate message that references the address we're hitting. Additionally, metadata is exchanged by default. The connect is used purely for it's effect of allowing this metadata exchange.

It also means that we will no longer need `is_local` to be provided, since this can be determined on the fly always. Providing `is_local` will override whatever is determined. By default this means that you're better off not providing `is_local` unless debugging this behaviour.

Reviewed By: c-ryan747

Differential Revision: D23476592

fbshipit-source-id: ea5032b617b6b805f53b5087a02c96c5dfde0433