ls supports multiple files on the backend
Summary:
`ls` now support listing of multiple files at the same time, bringing it inline with the other commands and `ls` more generally.

In this case we have to preserve compatibility with older clients sending a single path and populate this data. The data is populated for both, so newer clients can do the multi-listing

Reviewed By: c-ryan747

Differential Revision: D24078335

fbshipit-source-id: 2d378eb063c9b8ace7002099105eca3a5d89b0fd