Identical behaviour for logic test execution between booted and non-booted
Summary:
Logic tests can operate against non-booted sims. This is the approach that `fbxctest` takes.

However, in `idb` this will depend on whether the companion is running against a booted sim or not. This can lead to odd behaviours with audio units in iOS 14.

The best approach right now is to make the launching environment the same for both these cases. By doing so, we'll eliminate the subtle differences as a result.

Reviewed By: c-ryan747

Differential Revision: D24276645

fbshipit-source-id: 9db126bff1b3d9204155bb15d9d83c2b38422188