Optionally avoid fetching process state
Summary:
This is useful if you only care about whether is installed rather than installed and running.

For devices this can be quicker since there's no dev image mounting.

We need to have the default case in the proto mean "please fetch" as this is the default for older clients that aren't yet providing this information.

Reviewed By: c-ryan747

Differential Revision: D24358710

fbshipit-source-id: 9b7498b9264ea7f1d1a030b7e5c5ecdef2e657c5