Detect secure connection to prevent user error
Summary:
In order to support services that start using Secure connections over different Xcode versions, it's very helpful to have this context part of the connection object that wraps the underlying connection. This prevents consumers of the transfer API from erroneously using insecure transfer on a secure connection. Doing so will lead to hangs and heap smashers:

- A connection is secure and the consumer is using insecure transfer
- A consumer reads from some protocol, the protocol may contain some value that indicates the read length. This is the case in instruments, where the read length is present in a struct returned in a read.
- An insecure read on a secure channel will lead to this value being nonsense, and very likely some high value instead of a very small read (which is more likely in these protocols). Either way this value as well as all data is garbage.
- The read size is then used in consumption. If it's some corrupt value this can lead to consuming this forever, or until we hit some heap overflow.

This isn't always used, there are many examples that use the Plist-wiring, which will automatically do the right thing in `MobileDevice.framework`, for cases that do raw transfer of bytes we can now ensure that the consumer of the socket doesn't need to be aware of whether the connection is secure or not and act accordingly. A connection will either be secure or insecure, there would be no reason to do both at the same time.

All current usages of "transfer objects", based upon the changes in this diff, there are only a handful of cases that we need to consider:
- Instruments (this is covered by list-apps)
- Crash report mover (covered by crash-list)
- Springboard (covered by describe with additional diagnostics)
- Logging (covered by idb log)

Everything else will be using higher level APIs, which accounts for this directly. Allowing the consumer of the API to specify if they use raw transfer or not is a bad idea, as these can change in iOS versions. We want to avoid hitting changes in connection security in iOS versions that can cause this.

This then fixes the issue on specific versions before iOS 14 where instruments would stop working due to confusion over insecure wiriing.

Reviewed By: Amadman, c-ryan747

Differential Revision: D24416187

fbshipit-source-id: ab4edf790e9f4d2d5c45d8f454182d9e7efbdc14