Use FBDataConsumerStackConsuming as a marker of stack-consumption
Summary:
This new restriction is not yet used in any API, but it is useful for users of a consumer who wish to pass in data that is only valid for the life of a stack frame.

Buffers that are passed into a `NSData` object that are only valid on the stack (i.e. alloca or a stack-buffer) will not see their buffers being copied automatically with an `-[NSData copy]` call. However, using these buffers is a great optimization when an external API exposes them.

For example in video streaming, we can end up re-using buffers internal to `AVFoundation`. Instead of copying these internal buffers, we wish to pass them to a consumer wrapped in an `NSData` object. However, a problem arises when passing this to a consumer of data that does not stop using the data once `consumeData:` returns. This is the case in a block-based consumer on a background queue; the data is used asynchronously with the caller, after which time the stack-allocated buffer may now be garbage.

To guard against this situation, we can use protocol membership as a way of an API telling the caller that the consumer passed into it must *not ever use the buffer of the NSData that is passed into it after returning*. This prevents a use-after-free scenario.

Reviewed By: c-ryan747

Differential Revision: D24503394

fbshipit-source-id: b059945a6d9afb3cfb8c54155fe07460816925aa