No-copy NAL writes on H264 Stream
Summary:
Now that we have good assumptions about the behaviour of the caller in relation to it's usage of Data, we can now pass data that is only valid in the caller's stack frame scopy to the consumer.

This means that there is far less copying of data within the pipeline as we re-use the same underlying buffers.

Reviewed By: c-ryan747

Differential Revision: D24503506

fbshipit-source-id: 325fd58c4367ac47f70cba2396ad7e395c54fd0b