Add MJPEG for Devices
Summary:
Implementing MJPEG is pretty easy since there is support for JPEGs in the encoder, and the wire format is just a stream of JPEGs itself.

This is vastly more efficient when compared to using CoreImage with RGBA in-memory buffers. Here we get a JPEG back in the buffer itself, so which results in using substantially less memory overall.

To do this we just need to configure the video output correctly and then enumerate the block buffer that we get from it.

Reviewed By: c-ryan747

Differential Revision: D24475330

fbshipit-source-id: 65aebe8fc73d096e94a2bd0637e0e92552be3f05