Fix signed integer conversions in incorrect 32-to-64 bit cast in AMDServiceConnections.
Summary:
I've found out that `AMDServiceConnectionReceive` and `AMDServiceConnectionSend` are both returning a value that is not `ssize_t` wide.

The value returned from these is actually a `int32`, which means that when cast to an `ssize_t` a return value of -1 comes out as 4294967295 (which is 2^32 - 1). This is an indicator that this is a signed 32 bit integer being cast to a 64-bit integer without conversion of the signed bit.

We now account for this and also the presence of `errno` in the return of these functions. As a result we won't crash on a broken pipe (we don't read 32bit_INTMAX).

Reviewed By: Amadman

Differential Revision: D24565142

fbshipit-source-id: db0d2b59c89d9f2d84a82ebebaa5bd13304aadf8