pass env variables to xcodebuild for method swizzling
Summary:
When `SIM_DEVICE_SET_PATH` is provided, we shall pass it to xcodebuild. And we also need to update `DYLD_INSERT_LIBRARIES` to include libMaculator.dylib that includes the codes for xcodebuild method swizzling.

`shims` is added to `FBTestLaunchConfiguration` as an optional property. It's first resolved in `FBXCTestRunRequest_AppTest` (in a similar way as in `FBXCTestRunRequest_LogicTest`), and then passed to `FBTestLaunchConfiguration`. Finally, `FBXcodeBuildOperation` tries to read the property if `SIM_DEVICE_SET_PATH` is provided.

Reviewed By: lawrencelomax

Differential Revision: D24560202

fbshipit-source-id: 9949326082c5b953501bb8b1f47e4f8b888d700f