convert HEIC screenshots to JPEG if applicable
Summary:
Apple tries to encode screenshots generated in UI tests with HEIC instead of JPEG starting in Xcode 12: https://developer.apple.com/documentation/xcode-release-notes/xcode-12-release-notes

So far, it seems like screenshots for iPhone 7/8 tests are encoded with HEIC whereas those for iPhone 6 are still encoded as JPEG. HEIC is not well supported on the Python side. So when we extract screenshots from the result bundle, convert them to JPEGs if they are encoded in HEIC. Throw an error if other unexpected encodings are used.

Reviewed By: nslingerland

Differential Revision: D24850331

fbshipit-source-id: 94d4c1906424d5fb07a9a67993c49f9a1224afd8