Pass the logs from stdout for each test to the test result
Summary:
When an XCTest emits stdout during a test, it is currently captured by idb, and printed as part of the entire log stream. However, that data is not associated with a specific test. You can see this by running idb with the `--json` flag and seeing that the log parameter returns [] for each test.

This diff adds basic support to parse the stream of logs and to associate the lines with each test. It does so by looking for the begin-test and end-test markers in the stdout to know which tests created the lines

Reviewed By: lawrencelomax

Differential Revision: D24781206

fbshipit-source-id: 85c887be7d5d2373515b1eb0333f573db341135a