fix screenshot extraction for Xcode 12.2
Summary:
Starting from Xcode 12, Apple will try to encode the screenshots with HEIC:
* In Xcode 12.0, screenshots for iPhone 7/8 tests are encoded with HEIC but the those screenshots still have a `.jpg` extension.
* In Xcode 12.2, Apple fixed the HEIC screenshots to have a `.heic` extension.

What we do here is: before calling `sips` to convert the HEIC screenshot, make sure the exported screenshot file to always has a `.jpg` extension. And then call `sips` to convert the format from HEIC to JPEG if it's encoded with HEIC. The input and output of `sips` are the same so the original sceenshot will be replaced with the new converted one. This way, we do not need to remove the origial one.

Reviewed By: nslingerland

Differential Revision: D24965879

fbshipit-source-id: 993f7d2664a98c48c86873f968025103b043c4e4