Use invocation to ensure consistency of class naming between listing and executing
Summary:
We're currently omitting the Swift module name when running test suites and individual tests. This means that the results between the shim for listing and for execution are inconsistent. This is largely due to the fact we rely on internal XCTest methods that strip the module name unlike the listing.

In order to fix this, the way that we parse test names out has been updated to introspect the invocation and then extract the class name, the reconstruct the test name from this. We also have to do this with test suites so that if a suite is a class, then the class also has the swift module notation.

There are no changes for Objc only tests, since there is no module notation in this case.

Reviewed By: c-ryan747

Differential Revision: D24954221

fbshipit-source-id: 01f6d2691417827848c8bfa78a4f7994168c530e