add placeholder support to xctestrun file
Summary:
For simulator tests, we cannot use the pre-installed artifacts on the device (`UseDestinationArtifacts` is false). Instead, we need to provide `UITargetAppPath` and `TestHostPath`. Currently, IDB client will copy the app located at `TestHostPath` over to the companion host, or else it will report an error. For MobileLab simulator tests, it's not necessary as it might already exist on the companion host. IDB client does not know the location though, but we can put placeholders there.

The reasons why I put a placeholder instead of leaving it empty and asking IDB companion to figure it out are:
1, `TestHostPath` and `UITargetAppPath` are required fields when `UseDestinationArtifacts` is false according to the xctestrun file standard. It's better to follow it.
2, We still need to provide the target app and test runner app info especially when multiple apps are there. IDB companion could still possible figure out the right test runner and target apps based on the xctestrun contents, but it's not as reliable as we specifying them explicitly.

Reviewed By: nslingerland

Differential Revision: D25018465

fbshipit-source-id: c80401b16876f8a59055e9de39600792df9e6ea7