Remove duplicate test naming swizzling
Summary:
This is not possible in `xctest` bundles as we use both the class and method name. You will have a compiler error if the same class has the same test method.

I've also proven that this cannot happen in subclasses also. (see test plan). The time that this fix was applied pre-fork shows that this relates to OCUnit and SenTest, which are not supported by this shim having been fully replaced by `XCTest` instead.

Reviewed By: c-ryan747

Differential Revision: D25120151

fbshipit-source-id: ce91da454daab1188b685d2469fbe305c1e1b455