Add JSON output for test listing
Summary: Instead of doing line by line reading, we can instead just dump the entire output of all tests and read that straight into a JSON parser. This makes the entire process a lot easier, and we'll now have full test details, even if we don't yet propagate it all.

Reviewed By: c-ryan747

Differential Revision: D25123049

fbshipit-source-id: 76b829caccd7152f5c0fb4bb8c4749216754b381