Less spammy application install checks
Summary:
Uses the underlying `CoreSimulator` method for determining if an Application is installed or not. Currently, install checking goes through a long path of parsing information about the existence of an installed application and then throwing it away.

Instead, we can determine precisely if the app is installed before attempting to fetch it's details. This will also save on confusing errors messaging where we log that an app is not installed, when the intent is to determine whether it is installed or not.

Reviewed By: zeyadsalloum

Differential Revision: D25274009

fbshipit-source-id: 37e66c3f28e822a9017875563dda2e02c46f7fd8