create a reader to read and expand xctestrun file
Summary:
We need to read the contents of a xctestrun file in three cases: 1, Install a xctest with `idb xctest install wilde.xctestrun`; 2, Run a xctest; 3, list installed tests (Thanks Lawrence for pointing it out). There are some placeholders that need to be expanded. Create a reader to read and expand the contents of a xctestrun file:
1, It's a centralized way to read a xctestrun file. The contents could be used right after without worrying about placeholder expansion.
2, We can easily support new placeholders. `__IDB_APPSTORAGE__` is our own. Different from device tests, simulator tests do not use pre-installed artifacts on the device, instead, we need to provide the target app and test runner app locations on the host. `__IDB_APPSTORAGE__` expands to the IDB storage location for apps.

If there is an error during the process of reading the xctestrun file contents, we will pop the error up until it get wrapped in a future.

Reviewed By: lawrencelomax

Differential Revision: D25018481

fbshipit-source-id: 5578eae9e76dee3139aee414e170423dba5bfec3