Use process timeout instead of an erroneous shim timeout.
Summary:
This is not responded to by the shim and is therefore useless.

The `testTimeout` value is instead used during the execution of the process. This is preferable to do from the outside as it requires no intervention on the part of the shim and it is also combined with a "sample" of the process which will give accurate stack state for why the test has hung.

Reviewed By: zeyadsalloum

Differential Revision: D25119799

fbshipit-source-id: 6521a917543d58bacba7013270ab80f47eeb988d