Permit discovery of booted but unpaired devices
Summary:
There is a circumstance where we will ignore the `AMDevice` value for a booted device that is only ever then represented by an `AMRestorableDevice`.

This occurs because "pairing" is always validated when obtaining `AMDevice` values. This is undesirable as it prevents us from being able to pull out values like the UDID for a device that we otherwise could do. Instead, it makes more sense to fetch the values here and to flag the device as unpaired. Starting pairing with a session is still preferred as this provides more values than just a simple "Connect" call so we attempt to do so if possible. Performing a session w/ pairing means that we obtain information such as the backup info, amongst others

Because of the lack of pairing, operations later on will fail but in an informative way. There will be pairing failure messages in these eventualities which is vastly preferrable to the device "dissapearing"

Reviewed By: abesto

Differential Revision: D25494639

fbshipit-source-id: c21e20cc570170a889d45ef8dfde2edfd2757d35