Swallow plugin errors
Summary:
The idb plugin system is used to inject various loggers into core points of the idb client. As these are not essential for the call to succeed, we should swallow any exceptions produced by the plugins.

This will help isolate errors in logging from affecting the operation of idb's core functionality

Reviewed By: lawrencelomax

Differential Revision: D25756094

fbshipit-source-id: f8c6e032237929ba75481dee778dd4fbef25bfba