Remove FBVideoStreamAttributes
Summary:
This isn't needed anymore, it was only used for `fbsimctl`. Eventually we might need to be able to communicate the stream information out, however this should be in a structured format.

For now though, we can instead rely on video parsers being able to auto detect the video stream itself based upon headers and other information. For instance MJPEG stream sizes can be inferred from the size of the first JPEG image in the stream

Reviewed By: yancouto

Differential Revision: D25886106

fbshipit-source-id: 1798d200a2dfc6f3f74c7f812e91b708c626f870