Add CompressionQuality to FBVideoStreamConfiguration
Summary: Adds a compression quality value to the configuration. Also adds it to the idb grpc definition. Here the value is always defaulted to 0.2 if not provided. This is done because we can't guarantee that clients are sending this value yet. In time we will be able to remove the nullability in the constructor

Reviewed By: ideahitme

Differential Revision: D24724914

fbshipit-source-id: ec4aa892290f5f335cb247e6ae23af6ad0851592