Non-zero exit code when the test binary crashes outside a test case
Summary:
Currently when running a test idb always exits with exit code = 0, even when the underlying test binary has crashed.

In the case where the crash happens inside a test case usually wrappers of idb can handle this case, however when the crash happens outside of a test case it can be tricky to know what to report. Idb will output a result like
```
{
  "bundleName": "MyBundle",
  "className": "",
  "methodName": "",
  "logs": [],
  "duration": 0.0,
  "passed": false,
  "crashed": true,
  "failureInfo": {
    "message": "",
    "file": "",
    "line": 0
  },
  "activityLogs": []
}
```
which can get incorrectly reported as a test result.

To address this I'm making idb exit with code 3 when it detects that the test bundle has crashed outside of a test case.

Reviewed By: lawrencelomax

Differential Revision: D25995416

fbshipit-source-id: 6a106c1ca863d31b3fd679577268ebebc755eb5d