Show output of test listing on failure
Summary: In the same way that the error we print for running tests will show the output in the error message, we should do the same with test listing. At the same time, renaming the the "shim consumer" to "shim buffer" since this is a more concrete description of how it's used.

Reviewed By: c-ryan747

Differential Revision: D26075357

fbshipit-source-id: 4f3e00afa36e4781025f851e50a335bd9f349494