Orthogonalize FBLaunchedApplication from FBLaunchedProcess
Summary:
We've never had the ability to consistently get the exit status of a launched application in the same way as we can for launched binaries. This relates to the richness of the APIs available for doing so. For mac and iOS sim tasks we will always get the value that is returned in `waitpid(2)` which is a bitmask containing both the exit code and the signal. This is not true for applications.

As such, it is misleading to have an `exitCode` future that may well "faked". Instead of doing this, let's constrain the interface to the actual implementations and keep a separate interface for the process that we can expand in the future.

Reviewed By: c-ryan747

Differential Revision: D26099131

fbshipit-source-id: 5490800f8167f30e732e3df61acd28ec00aa57e5