Respect exit codes properly in FBTask
Summary:
The current behaviour of the `exitCode` future in FBTask isn't correct. We extract the signal from the mask and treat it as a exit code.

This expands the `FBTask` API to make `exitCode` now error if it was signalled. This does require some changes to tests, but error conditions tasks are by default exit != 0, so these will still fail but now in a better way identifying the signal instead of saying "this is the wrong exit and here is the wrong exit code"

Reviewed By: c-ryan747

Differential Revision: D26100163

fbshipit-source-id: 75c7504ac58f53a337bcb69fd705f9cf4ee97488