Respect exit codes properly in FBSimulatorAgentOperation
Summary:
The current behaviour of the `exitCode` future in `FBSimulatorAgentOperation` isn't correct. We extract the signal from the mask and treat it as a exit code.

This expands the `FBSimulatorAgentOperation` API to make `exitCode` now error if it was signalled. This does require some changes to tests, but error conditions tasks are by default exit != 0, so these will still fail but now in a better way identifying the signal instead of saying "this is the wrong exit and here is the wrong exit code"

Reviewed By: c-ryan747

Differential Revision: D26100887

fbshipit-source-id: c882cdbbedf375eeaaab8b91d30c26c4634d9da1