Push implementor's signal/statLoc/exitCode futures into FBLaunchedProcess
Summary:
Now that all of the implementors of this have the corresponding methods, we can generalise this to the protocol.

This also removes the ambiguity of the "completed" future in the protocol, so that callers must move towards the explicit exitCode etc. future to show the right intent.

`FBTask` is the exception here as it has `completed` future, now this is not overriding in the base protocol (as this is a new type). The `completed` future here accounts for status code checking

Reviewed By: c-ryan747

Differential Revision: D26099246

fbshipit-source-id: a40ceb690555c0eee425a5e019ac73633fd8fa56