Use only wedged process detection in FBListTestStrategy
Summary:
Crash logs take a while to populate and breaking down the behaviours of FBXCTestProcess into discrete methods allows for a mix-and-match approach.

Listing tests is supposed to be a short process and therefore is not streamed in idb. This should probably change but for the time being, let's just have the short-path of stalled process detection instead.

Reviewed By: c-ryan747

Differential Revision: D26127736

fbshipit-source-id: 87017b387adb36f3ce8b40668d2a0c95fc976f91