Wait for all process output to return before exiting launch
Summary:
When running `idb launch --wait-for XXX` theres a crash that can occur where we attempt to continue writing a process stdout + stderr after the grpc call has returned.

There are two potential fixes for this:
- Stop all writes after the grpc call has closed
- Wait for all writes before closing the grpc call

I'm going for the latter here as it might be useful to get the full output of a process on the client side. It does open up the opportunity for `idb launch` hanging if the launched process keeps running but this seems unlikely.

This is achieved by putting all of the writes in `FBBlockDataConsumer_Dispatcher` into a dispatch group. We then wait for that dispatch group to complete before returning in `consumeEndOfFile`. This changes the meaning of `consumer.finishedConsuming` to complete after the consumers have consumed all content, rather than just when they've been scheduled to. We can then await `consumer.finishedConsuming` before returning from `idb launch`

Reviewed By: lawrencelomax

Differential Revision: D26203254

fbshipit-source-id: 2d4ee6bf7b87edc94f35466f93f65e145986c267