Add Locale Set/Get/List to Companion
Summary:
Adds the ability to set/get/list locale identifiers for an iOS Simulator. This means a number of new APIs.

The methodology here is to use the defaults commands directly rather than using plists. This is vastly preferable as it doesn't require reading and writing of files. The listing is convenient as it means that it's possible to understand what is supported before sending a request.

Reviewed By: c-ryan747

Differential Revision: D25974670

fbshipit-source-id: 7e9320f931237e90a1dcbe1ce155a9ea2573a127