Fix precedence of App bundled dylib loading over System ones
Summary:
If an app embeds swift runtime overlays due to supporting earlier versions of iOS, then it may erroneously link those when running an Application Test.

`libSwiftCore.dylib` may be bundled inside `TestHostApp.app/Frameworks`. When loading other dependent frameworks these frameworks may pull in system level `libSwiftCore` and this will conflict with the version provided in the app bundle. If the Main App or Test Bundle depends on any Frameworks located within the Application bundle, then these should be referenced in the linker directives for these binaries.

We do still want to preserve the behaviour of allowing xctest targets linking against frameworks contained in the main app bundle. To do this, these are now moved to the set of fallback framework paths.

Reviewed By: c-ryan747

Differential Revision: D26401237

fbshipit-source-id: 870ef9a66995ac7408bcd10c73a37611b3af6108