Include last lines of output of stderr when logic test exits with a fatal code.
Summary:
When there is a non expected exit code, this is reasonably close to a crash but without a crash report generated. In these cases we want to avoid collecting a crash log, but the output of the `xctest` process is very significant.

What this change does it to accumulate stderr in a buffer so that it can be reported later on. This is typical of a linking problem that might occur.

Reviewed By: jbardini

Differential Revision: D26365733

fbshipit-source-id: 98f033b6dbe54926d4cc307f0a9056c5ad88342c