Constrain FBTestManager
Summary:
We don't need to expose the start to the test manager since it is wrapped upstream, the caller will always call `connect` as soon as it is returned. We can just do this in the static constructor.

This will allow for further changes where we can collapse a few redundant components into one and even `FBTestManager` itself

Reviewed By: c-ryan747

Differential Revision: D26426976

fbshipit-source-id: 8aecf8c4d15b295523aa513282af751e40f589f9