Remove FBTestDaemonResult
Summary: This class was needed before we had futures; there was a need for a return object that was essentially an `Either<T, Error>` via a class cluster. With Futures, we have this so stacking these objects on top of each other is very confusing and allows an error in the future but not the result and vice versa. Let's just move this to a single place where an error can occur; the future itself.

Reviewed By: c-ryan747

Differential Revision: D26450402

fbshipit-source-id: 5ff515db4e8d8f5793f0e58664291902215e534c