Override DYLD_FALLBACK_FRAMEWORK_PATH so list test workflow can find tests frameworks
Summary:
As title. Currently the list test workflow will load test bundle in a process using
the host app executable, and copy the test frameworks to app's Frameworks folder so test bundle
can be loaded. However, when actually running the test, the copied frameworks from developer library
are different than the one in iOS simulator's runtime, and not functioning properly.
Since we just need to be able to find those tests frameworks when loading test bundle in list test
workflow, we can use `DYLD_FALLBACK_FRAMEWORK_PATH` and add developer library frameworks' path to it.
In that way we can avoid hacking the app bundle directly.

Reviewed By: lawrencelomax

Differential Revision: D26453082

fbshipit-source-id: 3a531367f80e86dfea34a7ef319aec27f4742e59