Fix list target of app test bundle by passing app id to appPath parameter
Summary:
As title. Passing app path might not be very useful as idb_companion might be running
on a different machine than where idb is running. We can re-use appPath parameter to pass the
app bundle id so as long as we installed the app we can use its binary to list test targets.

Reviewed By: lawrencelomax

Differential Revision: D26453551

fbshipit-source-id: 58e32cf7024a343afd0438bcbed919a88fd9ddab