Flatten FBXCTestCommands API from Future<Future<Null>> to Future<Null>
Summary:
The ability to independently listen for execution started/ended is not significant. In all cases that I see in the usage, we only really care about the execution actually finishing overall.

This is highly beneficial as we can remove this nesting, make the API easier to deal with. Executing tests starts as soon as the future is returned. It is completed (successfully or otherwise) when the future resolves. Most of the time downstream the operation is already wrapped up in any case.

Reviewed By: c-ryan747

Differential Revision: D26459198

fbshipit-source-id: 9bcc43e17d0688498244b288b76b9fbc2e754d57