Flatten FBManagedTestRunStrategy
Summary:
Now that the consumer of this API has flattened it's two calls, the same can be done internally. This removes the need for any persisting of properties and self dances.

Callers now no longer need to do the flattening themselves either. Additionally, this will allow us to then flatten the mediator itself, closing further invariants within the API.

Reviewed By: c-ryan747

Differential Revision: D26459434

fbshipit-source-id: 3a2229ec3a1c1b26c20c9849181ec80d4f5ae94f