Flatten execution in FBTestManagerAPIMediator
Summary:
It's already consolidated in the callers, there's no need to expose this sequencing. Let's just push the behaviour into a single call that starts all of the execution. It will make future changes much easier that intend to remove more of the internals. Additionally, the need to handle additional invariants is now gone, since the API does now allow two executions to occur.

Cancellation now only needs handling in a single place, since there's no nesting of futures. This will open up even more changes to the internals and removal of state that was needed to apply invariants to the sequencing of the old public API.

Reviewed By: c-ryan747

Differential Revision: D26458319

fbshipit-source-id: 6799c21f4b529d16ff7c508d8b8cc35e78556d21