Report Error in Mediator as "Crash"
Summary:
Right now there is inadequate reporting to the reporter within idb's `xctest_run` when running application tests.

This is due to the fact that the companion handler depends on the reporter finishing as well as the operation ending. This means that we're replecating the behaviour of the logic test runner, which does handle errors identically to this.

There's an argument to be made that `xctest_run` in the companion should bail out when there is this kind of failure, however this keeps reporting in line with logic tests.

Reviewed By: jbardini

Differential Revision: D26514191

fbshipit-source-id: b636f0222a2562fe83ccbf577a4bd88bff0db6be