Pass down environment to child xctest process
Summary:
Now that we have a way of getting the in-process environment whilst removing env vars that could be troublesome, we now have a way of passing this down to logic and application tests. Logic Tests have a slight tweak so that parent env vars are override rather than override any clashing variables.

This can be done using the new API created in the parent diff.

Reviewed By: sandeepkumarpani888

Differential Revision: D26607256

fbshipit-source-id: 80c53af0161f0a85e112f9a1ade10446694dc0df