catch SystemExit in gen_main()
Summary: `SystemExit` does not inherit from `Exception` so we need to catch it explicitly. Or else the client will exit with an error code in case of `SystemExit`. `SystemExit` will be thrown by `argparse` when the arguments to `argparse` are not provided properly. P177293838 is one example of the error traceback. We need to catch `SystemExit` first before we can add tests to mutually exclusive arguments.

Reviewed By: c-ryan747

Differential Revision: D26505362

fbshipit-source-id: d9ff809b1db18235e5dfe29f57e1518136d847ba