Actually pass down timeout over grpc
Summary:
Before, this change, the timeout argument could be provided but it would not do anything.

The reason for this is that it just wasn't passed in the grpc. This meant that up until now the default value of 0 was always recieved in the companion. As such we retain this behaviour when `--timeout` is not provided. Passing `None` for the timeout will result in `0` being sent.

Reviewed By: jbardini

Differential Revision: D26636530

fbshipit-source-id: 97803c0b1a9566703d33362c6401ad0e7eafe529