Fix global timeout racing with test timeout
Summary:
Due to the way that cascading/nested timeouts are handled, we're not able to get a stackshot. The default for logic tests if 500s and the additional handling of stack sampling will take additional time on top.

As a result the stackshot will run when the entire execution times out. Just adjust the global timeout upwards to account for this

Reviewed By: c-ryan747

Differential Revision: D26637069

fbshipit-source-id: cdabafc9518c048a8e3e187a68aaf365fd576319